+++ t-sparqlUpdate/src/main/java/eu/unifiedviews/plugins/transformer/sparql/update/SparqlUpdate.java
@@ -1,34 +1,16 @@
 package eu.unifiedviews.plugins.transformer.sparql.update;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Date;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.regex.Pattern;
-
-import org.openrdf.model.URI;
-import org.openrdf.query.MalformedQueryException;
-import org.openrdf.query.QueryLanguage;
-import org.openrdf.query.Update;
-import org.openrdf.query.UpdateExecutionException;
-import org.openrdf.query.impl.DatasetImpl;
-import org.openrdf.repository.RepositoryConnection;
-import org.openrdf.repository.RepositoryException;
-
-import eu.unifiedviews.dpu.DPU;
-import eu.unifiedviews.dpu.DPUException;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import cz.cuni.mff.xrg.uv.transformer.sparql.update.SparqlUpdateConfig_V1;
 import eu.unifiedviews.dataunit.DataUnit;
 import eu.unifiedviews.dataunit.DataUnitException;
 import eu.unifiedviews.dataunit.rdf.RDFDataUnit;
 import eu.unifiedviews.dataunit.rdf.WritableRDFDataUnit;
+import eu.unifiedviews.dpu.DPU;
 import eu.unifiedviews.dpu.DPUContext;
+import eu.unifiedviews.dpu.DPUException;
 import eu.unifiedviews.helpers.dataunit.DataUnitUtils;
+import eu.unifiedviews.helpers.dataunit.copy.CopyHelper;
+import eu.unifiedviews.helpers.dataunit.copy.CopyHelpers;
 import eu.unifiedviews.helpers.dataunit.metadata.MetadataUtilsInstance;
 import eu.unifiedviews.helpers.dataunit.rdf.RdfDataUnitUtils;
 import eu.unifiedviews.helpers.dpu.config.ConfigHistory;
@@ -36,9 +18,21 @@ import eu.unifiedviews.helpers.dpu.config.migration.ConfigurationUpdate;
 import eu.unifiedviews.helpers.dpu.context.ContextUtils;
 import eu.unifiedviews.helpers.dpu.exec.AbstractDpu;
 import eu.unifiedviews.helpers.dpu.extension.ExtensionInitializer;
-import eu.unifiedviews.helpers.dpu.extension.faulttolerance.FaultTolerance;
 import eu.unifiedviews.helpers.dpu.extension.rdf.validation.RdfValidation;
 import eu.unifiedviews.plugins.transformer.sparql.SPARQLConfig_V1;
+import org.eclipse.rdf4j.model.IRI;
+import org.eclipse.rdf4j.query.MalformedQueryException;
+import org.eclipse.rdf4j.query.QueryLanguage;
+import org.eclipse.rdf4j.query.Update;
+import org.eclipse.rdf4j.query.UpdateExecutionException;
+import org.eclipse.rdf4j.query.impl.DatasetImpl;
+import org.eclipse.rdf4j.repository.RepositoryConnection;
+import org.eclipse.rdf4j.repository.RepositoryException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.*;
+import java.util.regex.Pattern;
 
 /**
  * @author Škoda Petr
@@ -65,9 +59,6 @@ public class SparqlUpdate extends AbstractDpu<SparqlUpdateConfig_V1> {
     @ExtensionInitializer.Init
     public RdfValidation rdfValidation;
 
-    @ExtensionInitializer.Init
-    public FaultTolerance faultTolerance;
-
     @ExtensionInitializer.Init(param = "eu.unifiedviews.plugins.transformer.sparql.SPARQLConfig__V1")
     public ConfigurationUpdate _ConfigurationUpdate;
 
@@ -78,75 +69,134 @@ public class SparqlUpdate extends AbstractDpu<SparqlUpdateConfig_V1> {
 
     @Override
     protected void innerExecute() throws DPUException {
-        if (useDataset()) {
-            ContextUtils.sendShortInfo(ctx, "sparqlUpdate.dpu.mode.openRdf");
+
+        boolean performanceOptimizationEnabled = false;
+
+        if (ctx.isPerformanceOptimizationEnabled(rdfInput)) {
+            performanceOptimizationEnabled = true;
+            if (!config.isPerGraph()) {
+                //activated, but no effect, as per graph option not selected
+                ContextUtils.sendInfo(ctx, "sparqlUpdate.dpu.opt.all.warn", "sparqlUpdate.dpu.opt.all.warn.detailed");
+            } else {
+                //activated
+                ContextUtils.sendInfo(ctx, "sparqlUpdate.dpu.optimistic.on", "sparqlUpdate.dpu.optimistic.on.detailed");
+            }
+
         } else {
-            ContextUtils.sendShortInfo(ctx, "sparqlUpdate.dpu.mode.virtuoso");
+            //not activated
+            ContextUtils.sendInfo(ctx, "sparqlUpdate.dpu.optimistic.off", "sparqlUpdate.dpu.optimistic.off.detailed");
+        }
+
+        if (useDataset()) {
+            //default and usual choice
+            LOG.debug("Dataset clause is used"); //ContextUtils.sendShortInfo(ctx, "sparqlUpdate.dpu.mode.openRdf");
+        } else {
+            LOG.debug("Dataset clause is NOT used"); //ContextUtils.sendShortInfo(ctx, "sparqlUpdate.dpu.mode.virtuoso");
         }
         // Get update query.
         final String query = config.getQuery();
         if (query == null || query.isEmpty()) {
             throw ContextUtils.dpuException(ctx, "sparqlUpdate.dpu.error.emptyQuery");
         }
-        final List<URI> outputGraphs = new LinkedList<>();
+        final List<IRI> outputGraphs = new LinkedList<>();
         // Get graphs.
         final List<RDFDataUnit.Entry> sourceEntries = getInputEntries(rdfInput);
-        // Copy data into a new graph.
+
+        //calculate input size (before executing changes)
+        long inputSize = 0;
+        try {
+            inputSize = getTriplesCount(rdfInput, RdfDataUnitUtils.asGraphs(sourceEntries));
+        } catch (DataUnitException e) {
+            LOG.error(e.getLocalizedMessage(), e.getStackTrace());
+        }
+
+        int numberOfInputGraphs = sourceEntries.size();
+        LOG.info("{} graph(s) in the input data unit", numberOfInputGraphs);
+
+        int numberOfOutputGraphs;
+
+
         if (config.isPerGraph()) {
             ContextUtils.sendMessage(ctx, DPUContext.MessageType.INFO, "sparqlUpdate.dpu.info.perGraph",
                     "sparqlUpdate.dpu.info.perGraph.body", sourceEntries.size());
-            // Execute on per-graph basis.
+            // Execute on per-graph basis. So if optimalized, it just copies the graph entries to the output data unit (with the original graphs),
+            // instead of producing new data graphs and copying content, and then executes on per-graph basis
             int counter = 1;
-            for (final RDFDataUnit.Entry sourceEntry : sourceEntries) {
-                LOG.info("Executing {}/{}", counter++, sourceEntries.size());
-                // For each input graph prepare output graph.
-                final URI targetGraph = faultTolerance.execute(new FaultTolerance.ActionReturn<URI>() {
+            CopyHelper copyHelper = CopyHelpers.create(rdfInput, rdfOutput);
+            try {
+                for (final RDFDataUnit.Entry sourceEntry : sourceEntries) {
+                    LOG.info("Executing {}/{}", counter++, sourceEntries.size());
 
-                    @Override
-                    public URI action() throws Exception {
-                        final URI outputUri = createOutputGraph(sourceEntry);
-                        LOG.info("   {} -> {}", sourceEntry.getDataGraphURI(), outputUri);
-                        return outputUri;
+                    String sourceSymbolicName;
+                    IRI sourceDataGraphURI;
+                    try {
+                        sourceSymbolicName = sourceEntry.getSymbolicName();
+                        sourceDataGraphURI = sourceEntry.getDataGraphURI();
+                    } catch (DataUnitException e) {
+                        throw new DPUException(e);
                     }
 
-                });
-                // Execute query 1 -> 1.
-                updateEntries(query, Arrays.asList(sourceEntry), targetGraph);
-                outputGraphs.add(targetGraph);
-                if (ctx.canceled()) {
-                    throw ContextUtils.dpuExceptionCancelled(ctx);
+                    //get target entries graph URL
+                    IRI targetGraph;
+                    if (performanceOptimizationEnabled) {
+                        try {
+                            copyHelper.copyMetadata(sourceSymbolicName);
+                        } catch (DataUnitException e) {
+                            throw new DPUException(e);
+                        }
+                        //target graph is the same as source graph
+                        try {
+                            targetGraph = sourceEntry.getDataGraphURI();
+                        } catch (DataUnitException e) {
+                            throw new DPUException(e);
+                        }
+
+                    }
+                    else {
+                        // For each input graph create new output graph
+                        targetGraph = createOutputGraph(sourceEntry);
+                    }
+
+                    LOG.info("Source: {}, graph {} is copied to graph {}", sourceSymbolicName, sourceDataGraphURI, targetGraph);
+
+                    // Execute query 1 -> 1.
+                    updateEntries(query, getGraphUriList(Arrays.asList(sourceEntry)), targetGraph, performanceOptimizationEnabled);
+                    outputGraphs.add(targetGraph);
+
+                    if (ctx.canceled()) {
+                        throw ContextUtils.dpuExceptionCancelled(ctx);
+                    }
                 }
+            } finally {
+                copyHelper.close();
             }
+
+            numberOfOutputGraphs = numberOfInputGraphs;
         } else {
-            // All graph at once, just check size.
-            if (sourceEntries.size() > MAX_GRAPH_COUNT) {
+            // Execute query on top of all graphs at once.
+            // In case of NON-optimalized approach (the original one) it creates only one OUTPUT graph, but in case of optimalized version it produces more graphs!
+            // For Optimized mode, this mode is not available!
+            ContextUtils.sendInfo(ctx, "sparqlUpdate.dpu.info.singleOutput", "sparqlUpdate.dpu.info.singleOutput.detailed");
+
+            if (numberOfInputGraphs > MAX_GRAPH_COUNT) {
                 throw ContextUtils.dpuException(ctx, "sparqlUpdate.dpu.error.tooManyGraphs",
                         MAX_GRAPH_COUNT, sourceEntries.size());
             }
-            // Prepare single output graph.
-            final URI targetGraph = faultTolerance.execute(new FaultTolerance.ActionReturn<URI>() {
 
-                @Override
-                public URI action() throws Exception {
-                    return createOutputGraph();
-                }
-            });
+            final IRI targetGraph = createOutputGraph();
+            LOG.info("All source graphs are copied to target graph {}", targetGraph);
+
             // Execute over all intpu graph ie. m -> 1
-            ContextUtils.sendShortInfo(ctx, "sparqlUpdate.dpu.info.singleOutput");
-            updateEntries(query, sourceEntries, targetGraph);
+            updateEntries(query, getGraphUriList(sourceEntries), targetGraph, false);
             outputGraphs.add(targetGraph);
+
+            numberOfOutputGraphs = 1;
         }
-        // Summmary message.
-        long inputSize = getTriplesCount(rdfOutput, faultTolerance.execute(new FaultTolerance.ActionReturn<URI[]>() {
 
-            @Override
-            public URI[] action() throws Exception {
-                return RdfDataUnitUtils.asGraphs(sourceEntries);
-            }
-        }));
-        long outputSize = getTriplesCount(rdfOutput, outputGraphs.toArray(new URI[0]));
+        // Calculate output size
+        long outputSize = getTriplesCount(rdfOutput, outputGraphs.toArray(new IRI[0]));
 
-        ContextUtils.sendShortInfo(ctx, "sparqlUpdate.dpu.msg.report", inputSize, outputSize);
+        ContextUtils.sendInfo(ctx, "sparqlUpdate.dpu.msg.report", "sparqlUpdate.dpu.msg.report.detailed", inputSize, outputSize, numberOfInputGraphs, numberOfOutputGraphs);
     }
 
     /**
@@ -154,30 +204,41 @@ public class SparqlUpdate extends AbstractDpu<SparqlUpdateConfig_V1> {
      * or over updateQuery graphs.
      * 
      * @param updateQuery
-     * @param sourceEntries
+     * @param sourceGraphs
      * @param targetgraph
+     * @param optimalized
      * @throws DPUException
      */
-    protected void updateEntries(final String updateQuery, final List<RDFDataUnit.Entry> sourceEntries,
-            final URI targetgraph) throws DPUException {
-        faultTolerance.execute(rdfInput, new FaultTolerance.ConnectionAction() {
+    protected void updateEntries(final String updateQuery, final List<IRI> sourceGraphs,
+                                 final IRI targetgraph, boolean optimalized) throws DPUException {
+        RepositoryConnection connection = null;
+        try {
+            connection = rdfInput.getConnection();
 
-            @Override
-            public void action(RepositoryConnection connection) throws Exception {
-                // Copy data.
-                executeUpdateQuery(QUERY_COPY, toUriList(sourceEntries), targetgraph, connection);
+            if (optimalized) {
+                // Execute user query
+                executeUpdateQuery(updateQuery, sourceGraphs, targetgraph, connection);
             }
-
-        });
-        faultTolerance.execute(rdfInput, new FaultTolerance.ConnectionAction() {
-
-            @Override
-            public void action(RepositoryConnection connection) throws Exception {
+            else {
+                //copy the data before
+                executeUpdateQuery(QUERY_COPY, sourceGraphs, targetgraph, connection);
                 // Execute user query over new graph.
                 executeUpdateQuery(updateQuery, Arrays.asList(targetgraph), targetgraph, connection);
             }
 
-        });
+        } catch (DataUnitException e) {
+            LOG.error(e.getLocalizedMessage(), e.getStackTrace());
+        } finally {
+            if (connection != null) {
+                try {
+                    connection.close();
+                } catch (RepositoryException e) {
+                    LOG.warn("Cannot close connection", e);
+                }
+            }
+        }
+
+
     }
 
     /**
@@ -191,7 +252,7 @@ public class SparqlUpdate extends AbstractDpu<SparqlUpdateConfig_V1> {
      * @param connection
      * @throws eu.unifiedviews.dpu.DPUException
      */
-    protected void executeUpdateQuery(String query, List<URI> sourceGraphs, URI targetGraph,
+    protected void executeUpdateQuery(String query, List<IRI> sourceGraphs, IRI targetGraph,
             RepositoryConnection connection) throws DPUException {
         // Prepare query.
         if (!useDataset()) {
@@ -202,13 +263,13 @@ public class SparqlUpdate extends AbstractDpu<SparqlUpdateConfig_V1> {
             }
             query = query.replaceFirst("(?i)WHERE", prepareUsingClause(sourceGraphs) + "WHERE");
         }
-        LOG.debug("Query to execute: {}", query);
+        LOG.info("Query to execute: {}", query);
         try {
             // Execute query.
             final Update update = connection.prepareUpdate(QueryLanguage.SPARQL, query);
             if (useDataset()) {
                 final DatasetImpl dataset = new DatasetImpl();
-                for (URI graph : sourceGraphs) {
+                for (IRI graph : sourceGraphs) {
                     dataset.addDefaultGraph(graph);
                 }
                 dataset.addDefaultRemoveGraph(targetGraph);
@@ -224,10 +285,12 @@ public class SparqlUpdate extends AbstractDpu<SparqlUpdateConfig_V1> {
     }
 
     /**
+     * Creates new output graph. Symbolic name is completely new (using time stamp in suffix).
+     * Data graph IRI is generated.
      * @return New output graph.
      * @throws DPUException
      */
-    protected URI createOutputGraph() throws DPUException {
+    protected IRI createOutputGraph() throws DPUException {
         // Register new output graph
         final String symbolicName = "http://unifiedviews.eu/resource/sparql-construct/"
                 + Long.toString((new Date()).getTime());
@@ -239,11 +302,13 @@ public class SparqlUpdate extends AbstractDpu<SparqlUpdateConfig_V1> {
     }
 
     /**
-     * @param symbolicName
+     * Creates new output graph. Symbolic name is based on the input graph (but modifed with this DPU's id as suffix).
+     * Data graph IRI is generated.
+     * @param entry
      * @return New output graph.
      * @throws DPUException
      */
-    protected URI createOutputGraph(RDFDataUnit.Entry entry) throws DPUException {
+    protected IRI createOutputGraph(RDFDataUnit.Entry entry) throws DPUException {
         final String suffix = "/" + ctx.getExecMasterContext().getDpuContext().getDpuInstanceId().toString();
         try {
             return rdfOutput.addNewDataGraph(entry.getSymbolicName() + suffix);
@@ -258,7 +323,7 @@ public class SparqlUpdate extends AbstractDpu<SparqlUpdateConfig_V1> {
      * @param graph
      * @return
      */
-    protected String prepareWithClause(URI graph) {
+    protected String prepareWithClause(IRI graph) {
         final StringBuilder withClause = new StringBuilder();
         withClause.append("WITH <");
         withClause.append(graph.stringValue());
@@ -273,22 +338,17 @@ public class SparqlUpdate extends AbstractDpu<SparqlUpdateConfig_V1> {
      * @return
      * @throws DPUException
      */
-    protected List<URI> toUriList(final List<RDFDataUnit.Entry> entries) throws DPUException {
-        final List<URI> result = new ArrayList<>(entries.size());
-        faultTolerance.execute(new FaultTolerance.Action() {
+    protected List<IRI> getGraphUriList(final List<RDFDataUnit.Entry> entries) throws DPUException {
+        final List<IRI> result = new ArrayList<>(entries.size());
 
-            @Override
-            public void action() throws Exception {
-                for (RDFDataUnit.Entry entry : entries) {
-                    try {
-                        result.add(entry.getDataGraphURI());
-                    } catch (DataUnitException ex) {
-                        throw ContextUtils.dpuException(ctx, ex, "sparqlUpdate.dpu.error.dataUnit");
-                    }
-                }
+        for (RDFDataUnit.Entry entry : entries) {
+            try {
+                result.add(entry.getDataGraphURI());
+            } catch (DataUnitException ex) {
+                throw ContextUtils.dpuException(ctx, ex, "sparqlUpdate.dpu.error.dataUnit");
             }
+        }
 
-        });
         return result;
     }
 
@@ -297,9 +357,9 @@ public class SparqlUpdate extends AbstractDpu<SparqlUpdateConfig_V1> {
      * @return Using clause for SPARQL insert, based on input graphs.
      * @throws DPUException
      */
-    protected String prepareUsingClause(List<URI> uris) throws DPUException {
+    protected String prepareUsingClause(List<IRI> uris) throws DPUException {
         final StringBuilder usingClause = new StringBuilder();
-        for (URI uri : uris) {
+        for (IRI uri : uris) {
             usingClause.append("USING <");
             usingClause.append(uri.stringValue());
             usingClause.append("> \n");
@@ -313,13 +373,11 @@ public class SparqlUpdate extends AbstractDpu<SparqlUpdateConfig_V1> {
      * @throws DPUException
      */
     protected List<RDFDataUnit.Entry> getInputEntries(final RDFDataUnit dataUnit) throws DPUException {
-        return faultTolerance.execute(new FaultTolerance.ActionReturn<List<RDFDataUnit.Entry>>() {
-
-            @Override
-            public List<RDFDataUnit.Entry> action() throws Exception {
-                return DataUnitUtils.getEntries(dataUnit, RDFDataUnit.Entry.class);
-            }
-        });
+        try {
+            return  DataUnitUtils.getEntries(dataUnit, RDFDataUnit.Entry.class);
+        } catch (DataUnitException e) {
+            throw new DPUException(e);
+        }
     }
 
     protected final boolean useDataset() {
@@ -329,29 +387,37 @@ public class SparqlUpdate extends AbstractDpu<SparqlUpdateConfig_V1> {
 
     /**
      * @param dataUnit
-     * @param entries
+     * @param graphs
      * @return Number of triples in given entries.
      */
-    protected Long getTriplesCount(final RDFDataUnit dataUnit, final URI[] graphs)
+    protected Long getTriplesCount(final RDFDataUnit dataUnit, final IRI[] graphs)
             throws DPUException {
-        return faultTolerance.execute(new FaultTolerance.ActionReturn<Long>() {
 
-            @Override
-            public Long action() throws Exception {
-                RepositoryConnection connection = null;
-                Long size = 0l;
-                try {
-                    connection = dataUnit.getConnection();
-                    size = connection.size(graphs);
-                } finally {
-                    if (connection != null) {
-                        connection.close();
-                    }
-                }
-                return size;
+        RepositoryConnection connection = null;
+        Long size = 0l;
+        try {
+            try {
+                connection = dataUnit.getConnection();
+            } catch (DataUnitException e) {
+                throw new DPUException(e);
             }
-
-        });
+            try {
+                size = connection.size(graphs);
+            } catch (RepositoryException e) {
+                throw new DPUException(e);
+            }
+        } finally {
+            if (connection != null) {
+                try {
+                    connection.close();
+                } catch (RepositoryException e) {
+                    throw new DPUException(e);
+                }
+            }
+        }
+        return size;
     }
+
+
 
 }