+++ t-tabular/src/main/java/eu/unifiedviews/plugins/transformer/tabular/column/ValueGeneratorTyped.java
@@ -5,9 +5,9 @@ import cz.cuni.mff.xrg.uv.transformer.tabular.column.ValueGeneratorReplace;
 import java.util.List;
 import java.util.Map;
 
-import org.openrdf.model.URI;
-import org.openrdf.model.Value;
-import org.openrdf.model.ValueFactory;
+import org.eclipse.rdf4j.model.IRI;
+import org.eclipse.rdf4j.model.Value;
+import org.eclipse.rdf4j.model.ValueFactory;
 
 import eu.unifiedviews.plugins.transformer.tabular.parser.ParseFailed;
 
@@ -20,9 +20,9 @@ public class ValueGeneratorTyped extends ValueGeneratorReplace {
 
     private final String typeStr;
 
-    private URI typeUri;
+    private IRI typeUri;
 
-    public ValueGeneratorTyped(URI uri, String template, String typeStr) {
+    public ValueGeneratorTyped(IRI uri, String template, String typeStr) {
         super(uri, template);
         this.typeStr = typeStr;
     }
@@ -40,7 +40,7 @@ public class ValueGeneratorTyped extends ValueGeneratorReplace {
     @Override
     public void compile(Map<String, Integer> nameToIndex, ValueFactory valueFactory) throws ParseFailed {
         super.compile(nameToIndex, valueFactory);
-        typeUri = valueFactory.createURI(typeStr);
+        typeUri = valueFactory.createIRI(typeStr);
     }
 
 }