+++ e-filesDownload/CHANGELOG.md
@@ -1,9 +1,11 @@
 E-FilesDownload
 ----------
 
-v2.1.7
+v3.0.0
 ---
 * Added support for custom delay (in ms)  between the HTTP calls
+* Adjusted creation of Virtual Path (see the user documentation)
+* Updated to use UnifiedViews helpers 3.X with rdf4j support
 
 v2.1.6
 ---