+++ q-sparqlAsk/CHANGELOG.md
@@ -1,6 +1,16 @@
 Q-SparqlAsk
 ----------
 
+v2.0.0
+---
+* Updated to use UnifiedViews helpers 3.X with rdf4j support
+
+
+v1.1.0
+---
+* Support for [RDF Data Processing Performance Optimization](https://grips.semantic-web.at/display/UDDOC/RDF+Data+Processing+Performance+Optimization)
+* Fault tolerance support removed (not used anyway)
+
 v1.0.0
 ---
 * Initial version